+++ lib/sparql/algebra/expression/function_call/function_call.ex
@@ -1,18 +1,25 @@
 defmodule SPARQL.Algebra.FunctionCall do
+  @moduledoc """
+  Shared functions for builtin and extension function calls.
+  """
 
   alias __MODULE__
   alias SPARQL.Algebra.Expression
 
+  @doc """
+  Evaluates argument expressions.
+
+  An unbound variable is evaluated to the value `:error`.
+
+  """
   def evaluate_arguments(arguments, data, execution) do
     {:ok, Enum.map(arguments, &(evaluate_argument(&1, data, execution)))}
-    # TODO: 17.2.1 Invocation: Numeric arguments are promoted as necessary to fit the expected types for that function or operator.
-    # TODO: 17.2.1 Invocation: If any of these steps fails, the invocation generates an error. The effects of errors are defined in Filter Evaluation.
   end
 
   def evaluate_argument(variable, data, execution)
 
   def evaluate_argument(variable, %{solution: solution}, _) when is_binary(variable),
-    do: solution[variable]
+    do: Map.get(solution, variable, :error)
 
   def evaluate_argument(%FunctionCall.Builtin{} = function_call, data, execution),
     do: Expression.evaluate(function_call, data, execution)